refactor: split statement types into separate files
