+++ internal/pkg/flink/types/statement.go
@@ -8,112 +8,7 @@ import (
 
 	"github.com/confluentinc/cli/internal/pkg/flink/internal/utils"
 	"github.com/confluentinc/cli/internal/pkg/output"
-	cliutils "github.com/confluentinc/cli/internal/pkg/utils"
 )
-
-type StatementResults struct {
-	Headers []string
-	Rows    []StatementResultRow
-}
-
-func (s *StatementResults) GetHeaders() []string {
-	if s == nil {
-		return []string{}
-	}
-	return s.Headers
-}
-
-func (s *StatementResults) GetRows() []StatementResultRow {
-	if s == nil {
-		return []StatementResultRow{}
-	}
-	return s.Rows
-}
-
-type StatementResultRow struct {
-	Operation StatementResultOperation
-	Fields    []StatementResultField
-}
-
-func (r *StatementResultRow) GetRowKey() string {
-	rowKey := strings.Builder{}
-	for idx, field := range r.GetFields() {
-		rowKey.WriteString(field.ToString())
-		if idx != len(r.GetFields())-1 {
-			rowKey.WriteString("-")
-		}
-	}
-	return rowKey.String()
-}
-
-func (r *StatementResultRow) GetFields() []StatementResultField {
-	if r == nil {
-		var fields []StatementResultField
-		return fields
-	}
-	return r.Fields
-}
-
-const (
-	INSERT        StatementResultOperation = 0
-	UPDATE_BEFORE StatementResultOperation = 1
-	UPDATE_AFTER  StatementResultOperation = 2
-	DELETE        StatementResultOperation = 3
-)
-
-type StatementResultOperation float64
-
-func (s StatementResultOperation) IsInsertOperation() bool {
-	return s == INSERT || s == UPDATE_AFTER
-}
-
-func (s StatementResultOperation) String() string {
-	switch s {
-	case INSERT:
-		return "+I"
-	case UPDATE_BEFORE:
-		return "-U"
-	case UPDATE_AFTER:
-		return "+U"
-	case DELETE:
-		return "-D"
-	}
-	return ""
-}
-
-type MockStatementResult struct {
-	ResultSchema     flinkgatewayv1alpha1.SqlV1alpha1ResultSchema
-	StatementResults flinkgatewayv1alpha1.SqlV1alpha1StatementResult
-}
-
-type StatementError struct {
-	Message          string
-	HttpResponseCode int
-	FailureMessage   string
-	Usage            []string
-	Suggestion       string
-}
-
-func (e *StatementError) Error() string {
-	if e == nil {
-		return ""
-	}
-	errStr := "Error: no message"
-	if e.Message != "" {
-		errStr = fmt.Sprintf("Error: %s", e.Message)
-	}
-	if len(e.Usage) > 0 {
-		errStr += fmt.Sprintf("\nUsage: %s", cliutils.ArrayToCommaDelimitedString(e.Usage, "or"))
-	}
-	if e.Suggestion != "" {
-		errStr += fmt.Sprintf("\nSuggestion: %s", e.Suggestion)
-	}
-	if e.FailureMessage != "" {
-		errStr += fmt.Sprintf("\nError details: %s", e.FailureMessage)
-	}
-
-	return errStr
-}
 
 type PHASE string
